Implemented support in workflows to receive signals
Signals are a way for something external to SWF to trigger something in running
workflows. A typical use case might be some human intervention like cancelling
an order.

The documentation for sending signals to a workflow:
  http://docs.aws.amazon.com/amazonswf/latest/apireference/API_SignalWorkflowExecution.html

The documentation of the event attributes for receiving signals:
  http://docs.aws.amazon.com/amazonswf/latest/apireference/API_WorkflowExecutionSignaledEventAttributes.html

Receiving signals is fairly trivial, either the signal is in the event history
or not. I chose to return a promise for the action.receiveSignal() function,
mostly since my use case is to wait in the workflow until receiving a signal
indicating an external task by a human has been completed. However for other use
cases like checking if an order is cancelled, treating it as asynchronous might
be counter intuitive, in which case it should probably be changed.

One thing to bear in mind is that unlike tasks, its unlikely the signal sender
has the unique runId, so then the workflowId must be used instead. This means
coming up with predictable yet unique workflowIds is very important if intending
to use signals in a workflow.