Fix failing tests related to childWorkflows
The problem was a circular dependency, where Context depends on childWorkflow
which in turn, to create a new workflow needs to provide a new Context.

See http://selfcontained.us/2012/05/08/node-js-circular-dependencies/ for a good
explanation.

I chose the hacky method of requiring the Context module at runtime, since this
is probably the only action to ever need the Context module, so having Context
module injecting itself to each action seemed more complicated and pointless.