Implemented support to receive the same signal multiple times
One use case is a workflow that involves the same human interaction at multiple
points in the workflow, for example a QA of a photo both before and after some
process on the photo. This could be resolved with child workflows, however then
the external process responsible for sending the signal would need to understand
the rules for knowing the different workflowIds.

So instead implementing support to track the number of times a signal of a given
name has been sent, and the different input values sent each time should make
these types of workflows easier to implement.