Lower the timeout for workflow executions for integration tests
Especially for the ReceiveSignal tests where I had set the timout to 5 minutes,
we started to notice that when tests fail for whatever reason, the workflows
still running would cause problems for subsequent running of the tests.

I explicelty set the time to 5 seconds for all tests, rather than just  change
the two tests set to 5 minutes, since although the default is still fairly low
at 20 seconds, it might be raised in the future unwittingy causing this problem
again, and with a faster setup and teardown being worked on in PR#13 even 20
seconds might prove a little too high.