Add support for receiveSignal in local workflows
I want to test some workflows that happen to use receiveSignal, so having them
implemented in the local workflow will be really useful. This means introducing
a "state" of which workflows are currently running so signals can be added to
their context. For now this is a simple global hash exposed via storeContext.

To keep backwards compatibility, I kept workflowId optional for executeWorkflow
but it will be needed if you want to reference it with signalWorkflowExecution
unless you look up the source code for the default workflowId. I also chose to
only support a promise interface since that is really all we will ever use.