Purge the complete and failed items in the contextStore for local workflow
If the local workflow is used for more than just running tests, for example used
to stub SWF for development environment, then growing forever is more likely to
become an issue. So this commit is just to keep the store manageable. I ended up
writing a test for it since it is a little more complicated now. I considered
using https://www.npmjs.com/package/object-sizeof to test the purging, but
settled on just exposing the current state instead since the only reason I store
failed and completed tasks is for debugging anyway, so may as well have a method
for it.