Allow specifying custom env vars for lambda decider function (#32)
Although there should be no waiting IO activity in deciders like making network
requests, there are still some situations where some logic in the workflows is
dependant on the environment they are running in. As soflow takes responsibility
for deploying the decider lambda, when used in this way it would be convenient
to specify any environment variables that decider should have.

When running you own activity workers this is not an issue since the deployment
is not managed by soflow.