Remove Node 6 and babel
AWS no longer supports lambda on Node 6, and given that the current default is
Node 10, we may as well remove babel and support for node 8 as well. Having to
use babel created a lot of complexity which I hope I have now removed. I also
"fixed" how errors from lambda functions where sent to the workflows so things
are a bit more consistent.